renamed api.py to slack.py because api.py is ambiguous. slack.py contains API calls specific to slack.
